# docs/swagger.yaml
basePath: /v1
definitions:
...
  api.CreateTagRequest:
    properties:
      data_type:
        type: string
      description:
        type: string
      name:
        type: string
      unit:
        type: string
    type: object
...
  api.UpdateTagRequest:
    properties:
      data_type:
        type: string
      description:
        type: string
      name:
        type: string
      unit:
        type: string
    type: object
...
  api.UserInfo:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
    type: object
...
  models.Tag:
    properties:
      created_at:
        type: string
      data_type:
        type: string
      description:
        type: string
      device_id:
        type: string
      id:
        type: string
      name:
        type: string
      unit:
        type: string
    type: object
  response.ErrorPayload:
    properties:
      code:
        type: string
      details:
        items:
          $ref: '#/definitions/response.ValidationErrorDetail'
        type: array
      message:
        type: string
      timestamp:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorPayload'
      success:
        type: boolean
    type: object
  response.Meta:
    properties:
      pagination:
        allOf:
        - $ref: '#/definitions/response.Pagination'
        description: Chỉ xuất hiện khi có phân trang, nếu không có sẽ tự động ẩn đi
      timestamp:
        type: string
    type: object
  response.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.ValidationErrorDetail:
    properties:
      field:
        type: string
      issue:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Linh Community
    url: https://github.com/linhcommunity/nlscada-cloud
  description: Backend API for NL SCADA Cloud
  title: NL SCADA Center API
  version: "1.0"
paths:
...
  /sites/{siteID}/devices/{deviceID}/tags:
    get:
      parameters:
      - description: Site UUID
        in: path
        name: siteID
        required: true
        type: string
      - description: Device UUID
        in: path
        name: deviceID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Danh sách tag
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Tag'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Danh sách tag của thiết bị
      tags:
      - Tags
    post:
      consumes:
      - application/json
      parameters:
      - description: Site UUID
        in: path
        name: siteID
        required: true
        type: string
      - description: Device UUID
        in: path
        name: deviceID
        required: true
        type: string
      - description: Thông tin tag
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CreateTagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Tag đã tạo
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Tag'
              type: object
      security:
      - BearerAuth: []
      summary: Tạo tag mới
      tags:
      - Tags
...
  /sites/{siteID}/tags/{tagID}:
    delete:
      parameters:
      - description: Site UUID
        in: path
        name: siteID
        required: true
        type: string
      - description: Tag UUID
        in: path
        name: tagID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Xóa tag
      tags:
      - Tags
    put:
      consumes:
      - application/json
      parameters:
      - description: Site UUID
        in: path
        name: siteID
        required: true
        type: string
      - description: Tag UUID
        in: path
        name: tagID
        required: true
        type: string
      - description: Các trường cần cập nhật
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.UpdateTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Tag đã cập nhật
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Tag'
              type: object
      security:
      - BearerAuth: []
      summary: Cập nhật tag
      tags:
      - Tags
  /users:
    get:
      description: Trả về danh sách tài khoản đã xác thực email để có thể mời vào
        site
      produces:
      - application/json
      responses:
        "200":
          description: Danh sách user
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/api.UserInfo'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Danh sách user đã xác thực
      tags:
      - Users
  /users/me:
    get:
      description: Trả về thông tin tài khoản của user hiện đang đăng nhập
      produces:
      - application/json
      responses:
        "200":
          description: Thông tin cá nhân
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/api.UserInfo'
              type: object
        "401":
          description: Chưa xác thực
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Thông tin cá nhân
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: Sử dụng "Bearer <token>", nhưng cookie session_token cũng được chấp
      nhận khi gọi từ browser
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"